requirements.txt
streamlit>=1.28.0
anthropic>=0.8.0
python-docx>=1.0.0
PyPDF2>=3.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
requests>=2.31.0
pylatexenc>=3.0.0
responses>=0.23.0
tqdm>=4.66.0
